13A013 Update
Update includes further reductions in SQL calls for Eight-Thirty Theme
(personal blog) as well as optimizations with Tweet listings (60-second
updates)

+++ lib/globals.php
@@ -20,7 +20,7 @@
         }
 
         // Return the Numeric Value
-        return intval($rVal);
+        return floatval($rVal);
     }
 
     /**
